requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.21
djangorestframework==3.15.2
dnspython==2.6.1
pymongo==4.10.1
python-decouple==3.8
sqlparse==0.5.3
typing-extensions==4.5.0  
tzdata==2025.2

# Dependencias de IA
scikit-learn==1.0.2  # Modelos IA
joblib==1.2.0        # Serialización modelos
seaborn==0.12.2      # Visualización heatmaps
matplotlib==3.5.3    # Gráficos
pandas ==1.3.5